#16 move concat builtin macro to macro expander
